chore: trace_decoder::iface
